Renamed Listing.jsx and updated the component to show listings as cards
